+++ example/lib/main.dart
@@ -8,7 +8,6 @@ void main() {
   runApp(const MyApp());
 }
 
-
 class MyApp extends StatefulWidget {
   const MyApp({super.key});
 
@@ -24,6 +23,7 @@ class _MyAppState extends State<MyApp> {
   static const defaultTolerance = 0.6;
   static const defaultPrecision = 0.8;
   static const defaultBufferSize = 8196;
+  static const defaultA4Reference = 440.0;
 
   String note = "";
   int midiNote = 0;
@@ -42,6 +42,7 @@ class _MyAppState extends State<MyApp> {
   Uint8List? pcmData;
   List<double>? streamData;
   double pitchDeviation = 0;
+  double a4Reference = defaultA4Reference;
 
   @override
   void initState() {
@@ -49,7 +50,7 @@ class _MyAppState extends State<MyApp> {
   }
 
   void onStartPressed() async {
-    if(!isRecording) {
+    if (!isRecording) {
       try {
         await pitchDetector.startDetection();
 
@@ -59,10 +60,15 @@ class _MyAppState extends State<MyApp> {
         });
         debugPrint("[START] Is Recording: $isRecording");
 
-        pitchDetector.setParameters(toleranceCents: defaultTolerance, bufferSize: defaultBufferSize, sampleRate: defaultSampleRate, minPrecision: defaultPrecision);
+        pitchDetector.setParameters(
+          toleranceCents: defaultTolerance,
+          bufferSize: defaultBufferSize,
+          sampleRate: defaultSampleRate,
+          minPrecision: defaultPrecision,
+          a4Reference: defaultA4Reference,
+        );
 
         _pitchSubscription = pitchDetector.onPitchDetected.listen((data) async {
-
           final stream = await pitchDetector.getRawDataFromStream();
           final pcm = await pitchDetector.getRawPcmDataFromStream();
 
@@ -83,24 +89,24 @@ class _MyAppState extends State<MyApp> {
             bufferSize = data['bufferSize'] ?? defaultBufferSize;
             sampleRate = data['sampleRate'] ?? defaultSampleRate;
             minPrecision = data['minPrecision'] ?? defaultPrecision;
+            minPrecision = data['minPrecision'] ?? defaultPrecision;
+            a4Reference = data['a4Reference'] ?? defaultA4Reference;
 
             streamData = stream;
             // debugPrint("RAW DATA: $streamData");
 
             pcmData = pcm;
             // debugPrint("PCM DATA: $pcmData");
-
           });
         });
-
-      } catch(e) {
+      } catch (e) {
         debugPrint("Start Recording Error: $e");
       }
     }
   }
 
   void onStopPressed() async {
-    if(isRecording) {
+    if (isRecording) {
       try {
         await _pitchSubscription?.cancel();
         _pitchSubscription = null;
@@ -130,6 +136,7 @@ class _MyAppState extends State<MyApp> {
       bufferSize = defaultBufferSize;
       sampleRate = defaultSampleRate;
       minPrecision = defaultPrecision;
+      a4Reference = defaultA4Reference;
 
       isRecording = false;
       accuracy = 0;
@@ -145,28 +152,22 @@ class _MyAppState extends State<MyApp> {
 
   Widget _startButton(Size size) {
     return SizedBox(
-      height: size.height*0.1,
-      width: size.width*0.1,
+      height: size.height * 0.1,
+      width: size.width * 0.1,
       child: IconButton(
         onPressed: onStartPressed,
-        icon: Icon(
-          Icons.play_arrow_rounded,
-          color: Colors.black,
-        ),
+        icon: Icon(Icons.play_arrow_rounded, color: Colors.black),
       ),
     );
   }
 
   Widget _stopButton(Size size) {
     return SizedBox(
-      height: size.height*0.1,
-      width: size.width*0.1,
+      height: size.height * 0.1,
+      width: size.width * 0.1,
       child: IconButton(
         onPressed: onStopPressed,
-        icon: Icon(
-          Icons.stop_rounded,
-          color: Colors.black,
-        ),
+        icon: Icon(Icons.stop_rounded, color: Colors.black),
       ),
     );
   }
@@ -177,9 +178,7 @@ class _MyAppState extends State<MyApp> {
 
     return MaterialApp(
       home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Pitch Detection'),
-        ),
+        appBar: AppBar(title: const Text('Pitch Detection')),
         body: Center(
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
@@ -190,20 +189,48 @@ class _MyAppState extends State<MyApp> {
               Text("Note: $note", style: TextStyle(fontSize: 18)),
               Text("Octave: $octave", style: TextStyle(fontSize: 18)),
               Text("Midi Note: $midiNote", style: TextStyle(fontSize: 18)),
-              Text("Frequency: ${frequency.toStringAsFixed(2)} Hz", style: TextStyle(fontSize: 18)),
+              Text(
+                "Frequency: ${frequency.toStringAsFixed(2)} Hz",
+                style: TextStyle(fontSize: 18),
+              ),
               SizedBox(height: size.height * 0.02),
               Text("Accuracy: $accuracy%", style: TextStyle(fontSize: 18)),
-              Text("Pitch Deviation: $pitchDeviation", style: TextStyle(fontSize: 18)),
-              Text("Volume: ${volume.toStringAsFixed(2)}", style: TextStyle(fontSize: 18)),
-              Text("Volume from DbSF: ${volumeDbFS.toStringAsFixed(2)}", style: TextStyle(fontSize: 18)),
+              Text(
+                "Pitch Deviation: $pitchDeviation",
+                style: TextStyle(fontSize: 18),
+              ),
+              Text(
+                "Volume: ${volume.toStringAsFixed(2)}",
+                style: TextStyle(fontSize: 18),
+              ),
+              Text(
+                "Volume from DbSF: ${volumeDbFS.toStringAsFixed(2)}",
+                style: TextStyle(fontSize: 18),
+              ),
               SizedBox(height: size.height * 0.02),
-              Text("Tolerance: ${toleranceCents.toStringAsFixed(2)}", style: TextStyle(fontSize: 18)),
-              Text("Min Precision: ${minPrecision.toStringAsFixed(2)}", style: TextStyle(fontSize: 18)),
+              Text(
+                "Tolerance: ${toleranceCents.toStringAsFixed(2)}",
+                style: TextStyle(fontSize: 18),
+              ),
+              Text(
+                "Min Precision: ${minPrecision.toStringAsFixed(2)}",
+                style: TextStyle(fontSize: 18),
+              ),
               Text("BufferSize: $bufferSize", style: TextStyle(fontSize: 16)),
               Text("SampleRate: $sampleRate", style: TextStyle(fontSize: 16)),
+              Text(
+                "A4 Reference: $a4Reference",
+                style: TextStyle(fontSize: 16),
+              ),
               SizedBox(height: size.height * 0.02),
               Text("IsRecording: $isRecording", style: TextStyle(fontSize: 16)),
-              Text("OnPitch", style: TextStyle(fontSize: 20, color: isOnPitch ? Colors.green : Colors.transparent)),
+              Text(
+                "OnPitch",
+                style: TextStyle(
+                  fontSize: 20,
+                  color: isOnPitch ? Colors.green : Colors.transparent,
+                ),
+              ),
               SizedBox(height: size.height * 0.01),
               isRecording ? _stopButton(size) : _startButton(size),
             ],
@@ -213,8 +240,3 @@ class _MyAppState extends State<MyApp> {
     );
   }
 }
-
-
-
-
-