+++ test/flutter_pitch_detection_method_channel_test.dart
@@ -5,20 +5,20 @@ import 'package:flutter_pitch_detection/flutter_pitch_detection_method_channel.d
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  MethodChannelFlutterPitchDetection platform = MethodChannelFlutterPitchDetection();
+  MethodChannelFlutterPitchDetection platform =
+      MethodChannelFlutterPitchDetection();
   const MethodChannel channel = MethodChannel('flutter_pitch_detection');
 
   setUp(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
-      channel,
-      (MethodCall methodCall) async {
-        return '42';
-      },
-    );
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, (MethodCall methodCall) async {
+          return '42';
+        });
   });
 
   tearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
   });
 
   // test('getPlatformVersion', () async {